# src/swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Canaan Towers",
        "description": "Canaan Towers API",
        "termsOfService": "https://google.com",
        "contact": {
          "name": "Canaan Towers services Ltd",
          "url": "https://google.com",
          "email": ""
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "v1"
    },
    "host": "localhost:5000",
    "basePath": "",
    "tags": [
        {
            "name": "User",
            "description": "API for Users in the system"
        },
        {
            "name": "Quotes",
            "description": "API for Quote request in the system"
        }
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "schemes": [
      "http",
      "https"
    ],
    "paths": {
        "/users": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get all Users in system",
                "responses": {
                    "200": {
                      "description": "Success",
                      "content": {
                        "application/json": {
                          "schema": {
                            "$ref": "#/components/schemas/UserInfo"
                          }
                        }
                      }
                    }
                }
            },
            "post": {
                "tags": [
                    "User"
                ],
                "requestBody": {
                    "content": {
                      "application/json": {
                        "schema": {
                          "$ref": "#/components/schemas/UserRegistrationModel"
                        }
                      }
                    }
                },
                "responses": {
                    "200": {
                      "description": "Success",
                      "content": {
                        "application/json": {
                          "schema": {
                            "$ref": "#/components/schemas/UserInfo"
                          }
                        }
                      }
                    }
                },
                "summary": "User registers in system"
            }
        },
        "/users/login": {
            "post": {
                "tags": [
                    "User"
                ],
                "requestBody": {
                    "content": {
                      "application/json": {
                        "schema": {
                          "$ref": "#/components/schemas/UserLoginModel"
                        }
                      }
                    }
                },
                "responses": {
                    "200": {
                      "description": "Success",
                      "content": {
                        "application/json": {
                          "schema": {
                            "$ref": "#/components/schemas/AuthenticationResponse"
                          }
                        }
                      }
                    }
                }
            }
        },
        "/users/profile": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "User view profile",
                "responses": {
                    "200": {
                      "description": "Success",
                      "content": {
                        "application/json": {
                          "schema": {
                            "$ref": "#/components/schemas/UserInfo"
                          }
                        }
                      }
                    }
                }
            },
            "patch": {
                "tags": [
                    "User"
                ],
                "summary": "User Updates profile",
                "requestBody": {
                    "content": {
                      "application/json": {
                        "schema": {
                          "$ref": "#/components/schemas/UserRegistrationModel"
                        }
                      }
                    }
                },
                "responses": {
                    "200": {
                      "description": "Success",
                      "content": {
                        "application/json": {
                          "schema": {
                            "$ref": "#/components/schemas/UserInfo"
                          }
                        }
                      }
                    }
                }
            }
        },
        "/users/logout": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "User logouts profile",
                "responses": {
                    "200": {
                      "description": "Success",
                      "content": {
                        "application/json": {
                          "schema": {
                            "$ref": "#/components/schemas/UserInfo"
                          }
                        }
                      }
                    }
                }
            }
        },
        "/users/delete": {
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "User deletes profile",
                "requestBody": {
                      "content": {
                        "application/json": {
                          "schema": {
                            "$ref": "#/components/schemas/UserRegistrationModel"
                          }
                        }
                      }
                  },
                  "responses": {
                      "200": {
                        "description": "Success",
                        "content": {
                          "application/json": {
                            "schema": {
                              "$ref": "#/components/schemas/UserInfo"
                            }
                          }
                        }
                      }
                  },
                  "summary": "User registers in system"
            }
        },
        "/quotes": {
            "post": {
                "tags": [
                    "Quotes"
                ],
                "summary": "User makes quote request",
                "requestBody": {
                    "content": {
                      "application/json": {
                        "schema": {
                          "$ref": "#/components/schemas/QuoteRequestModel"
                        }
                      }
                    }
                },
                "responses": {
                    "200": {
                      "description": "Success",
                      "content": {
                        "application/json": {
                          "schema": {
                            "$ref": "#/components/schemas/QuoteResponse"
                          }
                        }
                      }
                    }
                }
            },
            "get": {
                "tags": [
                    "Quotes"
                ],
                "summary": "User views quotes",
                "responses": {
                    "200": {
                      "description": "Success",
                      "content": {
                        "application/json": {
                          "schema": {
                            "$ref": "#/components/schemas/QuoteResponse"
                          }
                        }
                      }
                    }
                }
            }
        },
        "/quotes/{quoteId}": {
          "get": {
            "tags": [
              "Quotes"
            ],
            "summary": "User views single quote",
            "parameters": [
              {
                "name": "quoteId",
                "in": "path",
                "required": true,
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            ],
            "requestBody": {
              "content": {
                "application/json": {
                  "schema": {
                    "type": "string",
                    "nullable": true
                  }
                }
              }
            },
            "responses": {
              "200": {
                "description": "Success",
                "content": {
                  "application/json": {
                    "schema": {
                      "$ref": "#/components/schemas/QuoteResponse"
                    }
                  }
                }
              }
            }
          },
          "patch": {
            "tags": [
              "Quotes"
            ],
            "summary": "User updates his quotes",
            "requestBody": {
                  "content": {
                    "application/json": {
                      "schema": {
                        "$ref": "#/components/schemas/QuoteRequestModel"
                      }
                    }
                  }
              },
            "responses": {
                "200": {
                  "description": "Success",
                  "content": {
                    "application/json": {
                      "schema": {
                        "$ref": "#/components/schemas/QuoteResponse"
                      }
                    }
                  }
                }
            }
          },
          "delete": {
            "tags": [
              "Quotes"
            ],
            "parameters": [
              {
                "name": "quoteId",
                "in": "path",
                "required": true,
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            ],
            "requestBody": {
              "content": {
                "application/json": {
                  "schema": {
                    "type": "string",
                    "nullable": true
                  }
                }
              }
            },
            "responses": {
              "200": {
                "description": "Success",
                "content": {
                  "application/json": {
                    "schema": {
                      "$ref": "#/components/schemas/QuoteResponse"
                    }
                  }
                }
              }
            }
          }
        }
    },
    "components": {
      "schemas": {
        "UserRegistrationModel": {
          "required": [
            "email",
            "firstname",
            "lastname",
            "password",
            "number"
          ],
          "type": "object",
          "properties": {
            "firstname": {
              "maxLength": 100,
              "minLength": 2,
              "type": "string"
            },
            "lastname": {
              "maxLength": 100,
              "minLength": 2,
              "type": "string"
            },
            "email": {
              "type": "string",
              "format": "email"
            },
            "number": {
              "maxLength": 14,
              "minLength": 14,
              "type": "string"
            },
            "password": {
              "type": "string"
            }
          },
          "additionalProperties": false
        },
        "UserInfo": {
          "type": "object",
          "properties": {
            "id": {
              "type": "string",
              "format": "uuid"
            },
            "firsname": {
              "type": "string",
              "nullable": true
            },
            "lastname": {
              "type": "string",
              "nullable": true
            },
            "email": {
              "type": "string",
              "nullable": true
            },
            "number": {
              "type": "string",
              "nullable": true
            }
          },
          "additionalProperties": false
        },
        "UserLoginModel": {
          "type": "object",
          "properties": {
            "email": {
              "type": "string",
              "nullable": true
            },
            "password": {
              "type": "string",
              "nullable": true
            }
          },
          "additionalProperties": false
        },
        "AuthToken": {
            "type": "object",
            "properties": {
              "token": {
                "type": "string"
              }
            },
            "additionalProperties": false
          },
        "AuthenticatedResponseUserModel": {
          "type": "object",
          "properties": {
            "id": {
              "type": "string",
              "format": "uuid"
            },
            "firstname": {
              "type": "string",
              "nullable": true
            },
            "lastname": {
              "type": "string",
              "nullable": true
            },
            "email": {
              "type": "string",
              "nullable": true
            },
            "number": {
              "type": "string",
              "nullable": true
            }
          },
          "additionalProperties": false
        },
        "AuthenticationResponse": {
          "type": "object",
          "properties": {
            "user": {
              "$ref": "#/components/schemas/AuthenticatedResponseUserModel"
            },
            "token": {
                "type": "string",
                "nullable": true
            }
          },
          "additionalProperties": false
        },
        "QuoteRequestModel": {
          "required": [
            "email",
            "number",
            "message"
          ],
          "type": "object",
          "properties": {
            "email": {
              "type": "string",
              "format": "email"
            },
            "number": {
              "maxLength": 14,
              "minLength": 14,
              "type": "string"
            },
            "message": {
              "type": "string"
            }
          },
          "additionalProperties": false
        },
        "QuoteResponse": {
          "type": "object",
          "properties": {
            "id": {
              "type": "string",
              "format": "uuid"
            },
            "email": {
              "type": "string",
              "nullable": true
            },
            "number": {
              "type": "string",
              "nullable": true
            },
            "message": {
              "type": "string",
              "nullable": true
            },
            "owner": {
              "type": "string",
              "format": "uuid"
            }
          },
          "additionalProperties": false
        }
      },
      "securitySchemes": {
        "Bearer": {
          "type": "http",
          "description": "Enter JWT Bearer token",
          "scheme": "bearer",
          "bearerFormat": "JWT"
        }
      }
    },
    "security": [
      {
        "Bearer": [ ]
      }
    ]
  }